utils libr. update

+++ v4.0/ps_utils/ps_utils.h
@@ -6,7 +6,7 @@
 /*   By: onigiz <onigiz@student.42istanbul.com.t    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/01 03:39:19 by onigiz            #+#    #+#             */
-/*   Updated: 2023/09/09 16:07:08 by onigiz           ###   ########.fr       */
+/*   Updated: 2023/09/09 21:02:55 by onigiz           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_stack
 
 void	fprint(char *str);
 int		flen(char *str);
-int		int_checker(int number);
+int		int_checker(long int number);
 int		fatoi(const char *str);
 void	free_all(t_stack *stacks);
 void	error_handle(t_stack *stacks);
